Implement theory of strlen and strnlen functions. Also implement basic c-string theory.
In the basic c-string theory some general declarations and definitions are made.
These declarations and definitions will be used in other theories about functions on c-strings (such as strlen, strcpy, strcat etc.).
In the strlen theory two funcitons are defined: strlen and strnlen. Some properties for these functions are proved.