Introduce theory for the strcmp function.
Only strcmp function is implemented: strncmp needs to be created separatelly.
A helper-function pairCmp has been introduced in order to provide simplier way for the strcmp definition.
Both, pairCmp and strcmp take a 'comparator' function as a first argument.
The comparator takes a pair of two items and return some integer.
No comparator definition is given in the theory.
Properties of the comparator are declared (implicitely or explicitely) in each particular as a part of assumptions.