Implement takeBuff function. Remove redundant lemmas about takeFullCString.
The difference between takeBuff and takeCString is in the result for non-c-string argument.
If the argument of the takeBuff is not a c-string it returns argument itself (not [] like in takeCString).
Some lemmas about takeFullCString and takeCString was absolutelly identical.
In order to eliminate the redundancy only one lemma about relationship between takeFullCString and takeCString has been left.
The same is true for the takeBuff function.