Introduce model for a strpbrk function.
The model provides correct behavior for correct c-strings.
If args are not c-strings the behavior is considered as not defined.
Note that there are some cases in which "real" (implemented in C language) strpbrk returns some value for incorrect arguments without failure. For example, for strings A = "abcd\0" and C = "c", where C is not-terminated string the "real" strpbrk returns "cd\0".
On the other hand, in terms of model the behavior for strings A and C is not defined and it returns None.